Developed exponoeter functionality

+++ src/exponometer/inc/max44009.h
@@ -3,5 +3,5 @@
 #include "stm8s.h"
 
 void max44009_init();
-int max44009_get_lux();
+uint16_t max44009_get_lux();
 #endif
